data-encoder/ecommerce/vectors/requirements.txt
sentence-transformers
torch
git+https://github.com/openai/CLIP.git
pillow